+++ src/main/java/org/openfact/pe/idm/TotalRepresentation.java
@@ -2,42 +2,43 @@ package org.openfact.pe.idm;
 
 import javax.validation.constraints.Digits;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 public class TotalRepresentation {
 
     @NotNull
     @Digits(integer = 10, fraction = 2)
-    private Long pagar;
+    private BigDecimal pagar;
 
     @NotNull
     @Digits(integer = 10, fraction = 2)
-    private Long otrosCargos;
+    private BigDecimal otrosCargos;
 
     @NotNull
     @Digits(integer = 10, fraction = 2)
-    private Long descuentoGlobal;
+    private BigDecimal descuentoGlobal;
 
-    public Long getPagar() {
+    public BigDecimal getPagar() {
         return pagar;
     }
 
-    public void setPagar(Long pagar) {
+    public void setPagar(BigDecimal pagar) {
         this.pagar = pagar;
     }
 
-    public Long getOtrosCargos() {
+    public BigDecimal getOtrosCargos() {
         return otrosCargos;
     }
 
-    public void setOtrosCargos(Long otrosCargos) {
+    public void setOtrosCargos(BigDecimal otrosCargos) {
         this.otrosCargos = otrosCargos;
     }
 
-    public Long getDescuentoGlobal() {
+    public BigDecimal getDescuentoGlobal() {
         return descuentoGlobal;
     }
 
-    public void setDescuentoGlobal(Long descuentoGlobal) {
+    public void setDescuentoGlobal(BigDecimal descuentoGlobal) {
         this.descuentoGlobal = descuentoGlobal;
     }
 }
